Explicit size_t cast in calloc call
git-svn-id: svn+ssh://svn.phoenixbox.net/svn/gat/trunk@53 ccb0ee8c-1eea-e311-9d83-0002a5492622

+++ main.c
@@ -140,7 +140,7 @@ int main(int argc, char *argv[]) {
 		return EXIT_FAILURE;
 	}
 	
-	TrayItem **pool = calloc(argc - 1, sizeof(TrayItem));
+	TrayItem **pool = calloc((size_t)argc - 1, sizeof(TrayItem));
 	for (int i = 1; i < argc; i++) {
 		/* Use the last argument as the device name */
 		TrayItem *item = malloc(sizeof(TrayItem));
